# storage/api-docs/v2-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brillio API",
        "description": "API Backend for Brillio Orientation System",
        "contact": {
            "email": "contact@brillio.africa"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Brillio API Server"
        }
    ],
    "paths": {
...
        "/api/v2/chat/conversations": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Liste les conversations de l'utilisateur",
                "operationId": "90fa560259b66dff27d8b9a47eb4e4e2",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des conversations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Crée une nouvelle conversation d'orientation",
                "operationId": "25713776d2d08602b0b47d76edfe598f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nouvelle discussion d'orientation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conversation créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Conversation créée"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Solde insuffisant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/chat/conversations/{id}": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Récupère les messages d'une conversation d'orientation",
                "operationId": "5439dbdba102149581869cdb38eb96d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre de messages à récupérer",
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages récupérés avec succès"
                    },
                    "404": {
                        "description": "Conversation non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chat"
                ],
                "summary": "Supprime une conversation d'orientation",
                "operationId": "3c56ef349051d76a899a82cada0b8b07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation supprimée avec succès"
                    },
                    "404": {
                        "description": "Conversation non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/chat/send": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Envoie un message à l'IA ou à un humain",
                "operationId": "a38cc70f67d389f0464a85309bdb7c4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Bonjour, j'ai besoin d'aide pour mon orientation."
                                    },
                                    "conversation_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse du chatbot ou confirmation d'envoi"
                    }
                }
            }
        },
        "/api/v2/chat/conversations/{id}/request-human": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Demande l'intervention d'un conseiller humain sur la conversation",
                "operationId": "a1073f44ce0b7451b36ce6c912a6ede2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Demande de support humain enregistrée"
                    },
                    "402": {
                        "description": "Solde insuffisant"
                    },
                    "404": {
                        "description": "Conversation non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/chat/conversations/{id}/cancel-human": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Annule une demande d'intervention humaine en attente",
                "operationId": "a39761e1de2072f5b6bfc75d2b92a547",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la conversation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Demande annulée avec succès. Retour à l'orientation IA."
                    },
                    "400": {
                        "description": "Impossible d'annuler car un conseiller a déjà pris en charge la conversation"
                    },
                    "404": {
                        "description": "Conversation non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}